requirements.txt
asgiref==3.5.0
attrs==21.4.0
click==8.0.4
cmake==3.22.2
coverage==6.3.2
dlib==19.23.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.0.3
iniconfig==1.1.1
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
numpy==1.22.2
opencv-python==4.5.5.62
packaging==21.3
pickle-mixin==1.0.2
Pillow==9.0.1
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pyparsing==3.0.7
pytest==7.0.1
pytest-cov==3.0.0
tomli==2.0.1
Werkzeug==2.0.3
